requirements.txt
six
mock
PyYAML
